refactor: extract population scale helper
